+++ src/gui/dlg_signature_detail.h
@@ -38,7 +38,7 @@ class DlgSignatureDetail : public QDialog, public Ui::SignatureDetail {
     Q_OBJECT
 
 public:
-	DlgSignatureDetail(const MessageDb &messageDb, int dmId,
+	DlgSignatureDetail(const MessageDb &messageDb, qint64 dmId,
 	    QWidget *parent = 0);
 	DlgSignatureDetail(const void *msgDER, size_t msgSize,
 	    const void *tstDER, size_t tstSize, QWidget *parent = 0);
@@ -106,7 +106,7 @@ private:
 	bool tstInfo(QString &oStr, QString &ouStr, QString &nStr,
 	    QString &cStr) const;
 
-	 QSize dSize;
+	QSize dSize;
 };
